# api/openapi.json
{
  "components" : {
    "schemas" : {
...
      "ApiResponseInventoryItemResponse" : {
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/InventoryItemResponse"
          },
          "error" : {
            "type" : "string"
          },
          "success" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
      "ApiResponseInventoryStatsResponse" : {
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/InventoryStatsResponse"
          },
          "error" : {
            "type" : "string"
          },
          "success" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "ApiResponseListInventoryItemResponse" : {
        "properties" : {
          "data" : {
            "items" : {
              "$ref" : "#/components/schemas/InventoryItemResponse"
            },
            "type" : "array"
          },
          "error" : {
            "type" : "string"
          },
          "success" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "ApiResponseListServiceExpenseResponse" : {
        "properties" : {
          "data" : {
            "items" : {
              "$ref" : "#/components/schemas/ServiceExpenseResponse"
            },
            "type" : "array"
          },
          "error" : {
            "type" : "string"
          },
          "success" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "ApiResponseServiceExpenseResponse" : {
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/ServiceExpenseResponse"
          },
          "error" : {
            "type" : "string"
          },
          "success" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "ApiResponseTenantVisibleInventoryResponse" : {
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/TenantVisibleInventoryResponse"
          },
          "error" : {
            "type" : "string"
          },
          "success" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "CreateInventoryItemRequest" : {
        "properties" : {
          "category" : {
            "enum" : [ "APPLIANCES", "FURNITURE", "HVAC", "LAUNDRY", "ELECTRONICS", "FIXTURES", "SAFETY", "OTHER" ],
            "type" : "string"
          },
          "currentCondition" : {
            "enum" : [ "EXCELLENT", "GOOD", "FAIR", "DAMAGED" ],
            "type" : "string"
          },
          "modelNumber" : {
            "type" : "string"
          },
          "name" : {
            "minLength" : 1,
            "type" : "string"
          },
          "nextServiceDate" : {
            "format" : "date",
            "type" : "string"
          },
          "notes" : {
            "type" : "string"
          },
          "propertyId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "purchaseDate" : {
            "format" : "date",
            "type" : "string"
          },
          "replacementValue" : {
            "type" : "number"
          },
          "scope" : {
            "enum" : [ "PROPERTY_SHARED", "UNIT_PRIVATE" ],
            "type" : "string"
          },
          "serialNumber" : {
            "type" : "string"
          },
          "status" : {
            "enum" : [ "AVAILABLE", "ASSIGNED", "SHARED", "SERVICE_DUE", "RETIRED" ],
            "type" : "string"
          },
          "unitId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "warrantyExpiresAt" : {
            "format" : "date",
            "type" : "string"
          }
        },
        "required" : [ "category", "currentCondition", "name", "propertyId", "replacementValue", "scope", "status" ],
        "type" : "object"
      },
...
      "InventoryItemResponse" : {
        "properties" : {
          "category" : {
            "type" : "string"
          },
          "condition" : {
            "type" : "string"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "icon" : {
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "image" : {
            "type" : "string"
          },
          "location" : {
            "type" : "string"
          },
          "modelNumber" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "nextService" : {
            "type" : "string"
          },
          "notes" : {
            "type" : "string"
          },
          "propertyId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "serial" : {
            "type" : "string"
          },
          "shared" : {
            "type" : "boolean"
          },
          "status" : {
            "type" : "string"
          },
          "unitId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "value" : {
            "type" : "number"
          }
        },
        "type" : "object"
      },
      "InventoryStatsResponse" : {
        "properties" : {
          "maintenanceDue" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalAssets" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalValuation" : {
            "type" : "number"
          },
          "unassigned" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "ServiceExpenseRequest" : {
        "properties" : {
          "amount" : {
            "type" : "number"
          },
          "description" : {
            "minLength" : 1,
            "type" : "string"
          },
          "nextServiceDate" : {
            "format" : "date",
            "type" : "string"
          },
          "serviceDate" : {
            "format" : "date",
            "type" : "string"
          },
          "vendorName" : {
            "minLength" : 1,
            "type" : "string"
          }
        },
        "required" : [ "amount", "description", "serviceDate", "vendorName" ],
        "type" : "object"
      },
      "ServiceExpenseResponse" : {
        "properties" : {
          "amount" : {
            "type" : "number"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "itemId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "nextServiceDate" : {
            "format" : "date",
            "type" : "string"
          },
          "propertyId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "recordedBy" : {
            "format" : "uuid",
            "type" : "string"
          },
          "serviceDate" : {
            "format" : "date",
            "type" : "string"
          },
          "vendorName" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "TenantVisibleInventoryResponse" : {
        "properties" : {
          "sharedItems" : {
            "items" : {
              "$ref" : "#/components/schemas/InventoryItemResponse"
            },
            "type" : "array"
          },
          "unitItems" : {
            "items" : {
              "$ref" : "#/components/schemas/InventoryItemResponse"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
...
      "UpdateInventoryItemRequest" : {
        "properties" : {
          "category" : {
            "enum" : [ "APPLIANCES", "FURNITURE", "HVAC", "LAUNDRY", "ELECTRONICS", "FIXTURES", "SAFETY", "OTHER" ],
            "type" : "string"
          },
          "currentCondition" : {
            "enum" : [ "EXCELLENT", "GOOD", "FAIR", "DAMAGED" ],
            "type" : "string"
          },
          "modelNumber" : {
            "type" : "string"
          },
          "name" : {
            "minLength" : 1,
            "type" : "string"
          },
          "nextServiceDate" : {
            "format" : "date",
            "type" : "string"
          },
          "notes" : {
            "type" : "string"
          },
          "purchaseDate" : {
            "format" : "date",
            "type" : "string"
          },
          "replacementValue" : {
            "type" : "number"
          },
          "scope" : {
            "enum" : [ "PROPERTY_SHARED", "UNIT_PRIVATE" ],
            "type" : "string"
          },
          "serialNumber" : {
            "type" : "string"
          },
          "status" : {
            "enum" : [ "AVAILABLE", "ASSIGNED", "SHARED", "SERVICE_DUE", "RETIRED" ],
            "type" : "string"
          },
          "unitId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "warrantyExpiresAt" : {
            "format" : "date",
            "type" : "string"
          }
        },
        "required" : [ "category", "currentCondition", "name", "replacementValue", "scope", "status" ],
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "description" : "JWT access token from `POST /api/v1/auth/login` or `POST /api/v1/auth/signup`",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "contact" : {
      "email" : "dev@tenantliving.com",
      "name" : "Tenant Living Team"
    },
    "description" : "## Overview\nBackend API for the Tenant Living property management system.\n\n## Authentication\n- Obtain **`accessToken`** and **`refreshToken`** from **`POST /api/v1/auth/login`** or **`POST /api/v1/auth/signup`**.\n- Call protected endpoints with header **`Authorization`** value **`Bearer`** followed by the access token.\n- Rotate access with **`POST /api/v1/auth/refresh`** (send refresh token in JSON body).\n- Introspect a JWT without calling a protected route: **`POST /api/v1/auth/validate`**.\n\n## Errors\nValidation and business errors use **`application/json`** body **`ApiError`** (see schema).\n\n## Roles\n| Role | Description |\n|------|-------------|\n| SUPER_ADMIN | Primary landlord |\n| ADMIN | Co-owner / manager |\n| PROPERTY_STAFF | Operations staff |\n| USER | Tenant (default for self-signup) |\n",
    "license" : {
      "name" : "Proprietary",
      "url" : "https://tenantliving.com"
    },
    "title" : "Tenant Living API",
    "version" : "1.0.0"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/inventory/items" : {
      "post" : {
        "operationId" : "createItem",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateInventoryItemRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseInventoryItemResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "inventory-controller" ]
      }
    },
    "/api/v1/inventory/items/{itemId}" : {
      "get" : {
        "operationId" : "getItem",
        "parameters" : [ {
          "in" : "path",
          "name" : "itemId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseInventoryItemResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "inventory-controller" ]
      },
      "put" : {
        "operationId" : "updateItem",
        "parameters" : [ {
          "in" : "path",
          "name" : "itemId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateInventoryItemRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseInventoryItemResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "inventory-controller" ]
      }
    },
    "/api/v1/inventory/items/{itemId}/service-expenses" : {
      "get" : {
        "operationId" : "listServiceExpenses",
        "parameters" : [ {
          "in" : "path",
          "name" : "itemId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListServiceExpenseResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "inventory-controller" ]
      },
      "post" : {
        "operationId" : "recordServiceExpense",
        "parameters" : [ {
          "in" : "path",
          "name" : "itemId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ServiceExpenseRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseServiceExpenseResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "inventory-controller" ]
      }
    },
...
    "/api/v1/inventory/my-visible-items" : {
      "get" : {
        "operationId" : "getTenantVisibleItems",
        "parameters" : [ {
          "in" : "query",
          "name" : "propertyId",
          "required" : false,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseTenantVisibleInventoryResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "inventory-controller" ]
      }
    },
    "/api/v1/inventory/properties/{propertyId}/items" : {
      "get" : {
        "operationId" : "listPropertyItems",
        "parameters" : [ {
          "in" : "path",
          "name" : "propertyId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "q",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "status",
          "required" : false,
          "schema" : {
            "enum" : [ "AVAILABLE", "ASSIGNED", "SHARED", "SERVICE_DUE", "RETIRED" ],
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "scope",
          "required" : false,
          "schema" : {
            "enum" : [ "PROPERTY_SHARED", "UNIT_PRIVATE" ],
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "serviceDueOnly",
          "required" : false,
          "schema" : {
            "type" : "boolean"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListInventoryItemResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "inventory-controller" ]
      }
    },
    "/api/v1/inventory/properties/{propertyId}/stats" : {
      "get" : {
        "operationId" : "getPropertyStats",
        "parameters" : [ {
          "in" : "path",
          "name" : "propertyId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseInventoryStatsResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "inventory-controller" ]
      }
    },
...
  },
  "servers" : [ {
    "description" : "Local (default dev port)",
    "url" : "http://localhost:8080"
  } ]
}